organized my products and customers routes using express Roouter and creating a seperate routes folder
